SqlCommand now has better error reporting if search terms are malformed, ie. double-quoted like ""this""

+++ Fmp360_JDBC/src/com/prosc/fmpjdbc/FmConnection.java
@@ -76,7 +76,7 @@ public class FmConnection implements Connection {
 			packageLogger.setLevel(Level.parse(logLevel));
 		}
 		log.log(Level.CONFIG, "Connecting to " + url);
-		fmVersion = Float.valueOf( properties.getProperty("fmversion", "7") ).floatValue();
+		fmVersion = Float.valueOf( properties.getProperty( "fmversion", "7" ) );
 		if( fmVersion >= 7 ) {
 			//requestHandler = new FmXmlRequest( getProtocol(), getHost(), "/fmi/xml/FMPXMLRESULT.xml", getPort(), getUsername(), getPassword() );
 			//recIdHandler = new FmXmlRequest( getProtocol(), getHost(), "/fmi/xml/FMPXMLRESULT.xml", getPort(), getUsername(), getPassword(), fmVersion );
